+++ packages/sources/tradermade-test/src/ws-utils.ts
@@ -1,0 +1,9 @@
+import { WebsocketTransportGenerics } from '@chainlink/external-adapter-framework/transports'
+import { WebsocketReverseMappingTransport } from '@chainlink/external-adapter-framework/transports/websocket'
+
+export class TraderMadeWebsocketReverseMappingTransport<
+  T extends WebsocketTransportGenerics,
+  K,
+> extends WebsocketReverseMappingTransport<T, K> {
+  apiKey = ''
+}